+++ Pandora/Paradigm/Primary/Transformer/Construction.hs
@@ -1,6 +1,4 @@
-{-# LANGUAGE UndecidableInstances #-}
-
-module Pandora.Paradigm.Primary.Transformer.Construction (Construction (..), deconstruct, coiterate, section) where
+module Pandora.Paradigm.Primary.Transformer.Construction where
 
 import Pandora.Core.Functor (type (:.), type (:=), type (|->), type (~>))
 import Pandora.Pattern.Category ((.), ($))
@@ -22,7 +20,6 @@ import Pandora.Pattern.Object.Semigroup (Semigroup ((+)))
 import Pandora.Pattern.Object.Ringoid ((*))
 import Pandora.Pattern.Object.Monoid (Monoid (zero))
 import Pandora.Paradigm.Schemes.TU (TU (TU), type (<:.>))
-import Pandora.Paradigm.Controlflow.Effect.Interpreted (Interpreted (Primary, run))
 
 data Construction t a = Construct a (t :. Construction t := a)
 
@@ -93,16 +90,3 @@ instance (Covariant t, Avoidable u) => Avoidable (u <:.> Construction t) where
 
 instance (Traversable t, Traversable u) => Traversable (u <:.> Construction t) where
 	TU g ->> f = TU <$> g ->>> f
-
--- Experimental wrapper for data structures that can behave like list comprehensions
-newtype Comprehension t a = Comprehension (t <:.> Construction t := a)
-
-instance Interpreted (Comprehension t) where
-	type Primary (Comprehension t) a = t <:.> Construction t := a
-	run (Comprehension x) = x
-
-instance Covariant (t <:.> Construction t) => Covariant (Comprehension t) where
-	f <$> Comprehension x = Comprehension $ f <$> x
-
-instance (forall a . Semigroup (t <:.> Construction t := a), Bindable t) => Bindable (Comprehension t) where
-	Comprehension (TU t) >>= f = Comprehension . TU $ t >>= \(Construct x xs) -> run $ run (f x) + run (Comprehension (TU xs) >>= f)