+++ Pandora/Paradigm/Structure/Ability.hs
@@ -7,3 +7,4 @@ import Pandora.Paradigm.Structure.Ability.Rotatable as Exports
 import Pandora.Paradigm.Structure.Ability.Insertable as Exports
 import Pandora.Paradigm.Structure.Ability.Focusable as Exports
 import Pandora.Paradigm.Structure.Ability.Nonempty as Exports
+import Pandora.Paradigm.Structure.Ability.Comprehension as Exports